Before delete activity create and delete methods
