+++ pyproject.toml
@@ -24,7 +24,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
-testpaths = ["src/yclientsapi/tests/unit/src"]
+testpaths = ["src/yclientsapi/tests/integration/src"]
 markers = [
     "staff",
     "service",
