+++ src/yclientsapi/exceptions.py
@@ -1,3 +1,6 @@
-class BaseError(Exception):
+from httpx import HTTPError
+
+
+class YclientsApiResponseError(HTTPError):
     def __init__(self, message: str):
         super().__init__(message)
