--- src/yclientsapi/schedule.py
@@ -1,2 +1,0 @@
-# https://api.yclients.com/api/v1/activity/{company_id}/search/
-# https://developers.yclients.com/ru/#tag/Gruppovye-sobytiya/operation/api.location.activity.search_d
