+++ src/yclientsapi/tests/integration/src/tests/test_record.py
@@ -2,7 +2,7 @@ import pytest
 
 from yclientsapi.main import YclientsAPI
 from yclientsapi.schema.record import RecordListResponse
-from yclientsapi.tests.unit.src.test_data.real.record import Parametrize
+from yclientsapi.tests.integration.src.test_data.record import Parametrize
 
 
 @pytest.mark.parametrize(
